feat(rfc): prepare response object
